src/lfx/pyproject.toml
[project]
name = "lfx"
version = "0.1.12"
description = "Langflow Executor - A lightweight CLI tool for executing and serving Langflow AI flows"
readme = "README.md"
authors = [
    { name = "Gabriel Luiz Freitas Almeida", email = "gabriel@langflow.org" }
]
requires-python = ">=3.10,<3.14"
dependencies = [
    "langchain-core>=0.3.66",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "pillow>=10.0.0",
    "fastapi>=0.115.13",
    "uvicorn>=0.34.3",
    "typer>=0.16.0",
    "platformdirs>=4.3.8",
    "aiofiles>=24.1.0",
    "typing-extensions>=4.14.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "httpx[http2]>=0.24.0",
    "aiofile>=3.8.0",
    "json-repair>=0.30.3",
    "docstring-parser>=0.16",
    "networkx>=3.4.2",
    "nanoid>=2.0.0",
    "cachetools>=5.5.2",
    "emoji>=2.14.1",
    "chardet>=5.2.0",
    "defusedxml>=0.7.1",
    "passlib>=1.7.4",
    "pydantic-settings>=2.10.1",
    "tomli>=2.2.1",
    "orjson>=3.10.15",
    "asyncer>=0.0.8",
    "structlog",
    "loguru>=0.7.3",
    "langchain~=0.3.23",
    "validators>=0.34.0",
    "i18nice>=0.16.0",
]

[project.scripts]
lfx = "lfx.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lfx"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
pydocstyle.convention = "google"
select = ["ALL"]
ignore = [
    "C90", # McCabe complexity
    "CPY", # Missing copyright
    "COM812", # Messes with the formatter
    "ERA", # Eradicate commented-out code
    "FIX002", # Line contains TODO
    "ISC001", # Messes with the formatter
    "PERF203", # Rarely useful
    "PLR09", # Too many something (arg, statements, etc)
    "RUF012", # Pydantic models are currently not well detected. See https://github.com/astral-sh/ruff/issues/13630
    "TD002", # Missing author in TODO
    "TD003", # Missing issue link in TODO
    "TRY301", # A bit too harsh (Abstract `raise` to an inner function)
    "D1", # Missing docstring in public package
    "S1",
    "PLC0415", # Inline imports
    # Rules that are TODOs
    "ANN",
]

# Preview rules that are not yet activated
external = ["RUF027"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "D1",
    "PLR2004",
    "S101",
    "SLF001",
]
"src/lfx/base/*" = [
    "SLF001",
]
"src/lfx/components/*" = [
    "SLF001",
]
"src/lfx/custom/*" = [
    "SLF001",
]
"src/lfx/graph/*" = [
    "SLF001",
]

[tool.ruff.lint.flake8-builtins]
builtins-allowed-modules = [ "io", "logging", "socket"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short --strict-markers --disable-warnings --color=yes"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow-running tests",
    "asyncio: Async tests"
]

[dependency-groups]
dev = [
    "asgi-lifespan>=2.1.0",
    "blockbuster>=1.5.25",
    "coverage>=7.9.2",
    "hypothesis>=6.136.3",
    "pytest>=8.4.1",
    "pytest-asyncio>=0.26.0",
    "ruff>=0.9.10",
]
